+++ framework/Database/Migration.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Framework\Database;
+
+abstract class Migration implements MigrationInterface
+{
+}