+++ framework/Database/MigrationInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Framework\Database;
+
+interface MigrationInterface
+{
+    /** This function is called to apply the migration */
+    public function up(): void;
+}