+++ framework/Database/Migrations/2023_12_29_085321_create_migrations_table.php
@@ -1,0 +1,18 @@
+<?php
+
+use Framework\Database\Database;
+use Framework\Database\Migration;
+
+return new class extends Migration
+{
+    public function up(): void
+    {
+        Database::query(
+            'CREATE TABLE migrations (' .
+            'id INT auto_increment,' .
+            'name VARCHAR(255) NOT NULL,' .
+            'PRIMARY KEY (id)' .
+            ') ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;'
+        );
+    }
+};